+++ Logging.hpp
@@ -15,7 +15,7 @@ class Logging
 {
 public:
     Logging(SCStudyInterfaceRef sc);
-    void LogMetrics(SCStudyInterfaceRef sc, const std::string& strategyName, const std::string& reportPath, const std::vector<std::pair<std::string, int>>& params, int studyId);
+    void LogMetrics(SCStudyInterfaceRef sc, const std::string& strategyName, const std::string& reportPath, const std::vector<std::pair<std::string, double>>& params, int studyId);
 
     static void WriteTradesData(SCStudyInterfaceRef sc, std::ofstream &log);
     static void WriteTradeStatisticsV2(SCStudyInterfaceRef sc, std::ofstream &log);
@@ -27,10 +27,10 @@ private:
     std::string getCurrentDllName(SCStudyInterfaceRef sc, int studyId);
     std::string GetParameterValueAsString(const SCInputRef &Input);
     std::vector<std::pair<std::string, std::string>> GetParameters(int lastInputIndex);
-    void WriteSummaryHeader(std::ofstream &log, const std::string &strategyName, const std::string &dllName, const std::vector<std::pair<std::string, int>> &params);
+    void WriteSummaryHeader(std::ofstream &log, const std::string &strategyName, const std::string &dllName, const std::vector<std::pair<std::string, double>> &params);
 
     nlohmann::json GetCustomStudyInformation(SCStudyInterfaceRef sc, int studyId);
-    nlohmann::json GetCombination(const std::vector<std::pair<std::string, int>> &params);
+    nlohmann::json GetCombination(const std::vector<std::pair<std::string, double>> &params);
     nlohmann::json GetTradesData(SCStudyInterfaceRef sc);
     nlohmann::json GetTradeStatistics(SCStudyInterfaceRef sc);
     nlohmann::json GetStudyParameters(SCStudyInterfaceRef sc, int studyId);
