+++ StrategyOptimizer.hpp
@@ -11,8 +11,8 @@
 
 namespace StrategyOptimizerHelpers
 {
-    void HandleStartEvent(SCStudyInterfaceRef sc, SCInputRef Input_ConfigFilePath, StrategyOptimizerConfig* config, std::vector<std::vector<int>>* combinations, ReplayState& replayState, int& ComboIndex);
-    void HandleResetEvent(SCStudyInterfaceRef sc, ReplayState& replayState, int& ComboIndex, StrategyOptimizerConfig* config, std::vector<std::vector<int>>* combinations, Logging* logging);
+    void HandleStartEvent(SCStudyInterfaceRef sc, SCInputRef Input_ConfigFilePath, StrategyOptimizerConfig* config, std::vector<std::vector<double>>* combinations, ReplayState& replayState, int& ComboIndex);
+    void HandleResetEvent(SCStudyInterfaceRef sc, ReplayState& replayState, int& ComboIndex, StrategyOptimizerConfig* config, std::vector<std::vector<double>>* combinations, Logging* logging);
 }
 
 #endif
